requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
numpy==1.24.3
pandas==2.0.3
opencv-python==4.8.1.78
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1
pypdf2==3.0.1
python-docx==1.1.0
aiofiles==23.2.1
pydantic==2.5.0
google-generativeai==0.3.1
transformers==4.35.2
sentence-transformers==2.2.2
langchain==0.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-dotenv==1.0.0
